fjerner gammel endringslogg kode
