+++ cypress/integration/diverse_spec.js
@@ -3,18 +3,6 @@ before('Start server', () => {
 });
 
 describe('Diverse', () => {
-    //TODO toggle denne når tvungen stepper er av/på
-    // before('Tvungen stepper', () => {
-    //     cy.getByTestId('endringslogg_tour-modal').should('be.visible');
-    //     cy.getByTestId('endringslogg_neste-knapp')
-    //         .click();
-    //     cy.getByTestId('endringslogg_neste-knapp')
-    //         .click();
-    //     cy.getByTestId('endringslogg_ferdig-knapp')
-    //         .click();
-    //     cy.getByTestId('endringslogg_tour-modal').should('not.exist');
-    // });
-
     it('Verifiser tilbakemeldingsundersøkelse', () => {
         cy.getByTestId('tilbakemelding_modal').should('not.exist');
         cy.getByTestId('tilbakemelding_fab_knapp')
