--- src/components/modal/tour-modal/tour-modal.less
@@ -1,61 +1,0 @@
-@import '~nav-frontend-core/less/_variabler';
-
-.tour-modal {
-    padding: 0 !important;
-    background-color: @white;
-    display: flex;
-    flex-direction: column;
-    max-width: 31.25rem;
-    border-radius: @border-radius-large;
-
-    .lukknapp {
-        top: 1.25rem;
-        right: 1.5rem;
-    }
-
-    &__header {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-        padding: 1.125rem 0;
-        position: relative;
-        box-shadow: 0 2px 5px rgba(120, 112, 106, 0.5);
-    }
-
-    &__main {
-        &--bilde {
-            width: 100%;
-            height: 100%;
-        }
-
-        &--bilde-wrapper {
-            height: 12.5rem;
-            box-shadow: 0 2px 5px rgba(120, 112, 106, 0.5);
-        }
-
-        &--tekst {
-            margin: 0 2rem !important;
-            text-align: center;
-            min-height: 4.1rem;
-            max-width: 100%;
-        }
-
-        &--beskrivelse {
-            display: flex;
-            flex-direction: column;
-            justify-content: flex-end;
-            align-items: center;
-            margin: 1rem 0;
-            height: 8rem;
-        }
-    }
-
-    &__footer {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        margin: 2.5rem 1.5rem 1rem;
-    }
-}
