+++ src/routes.tsx
@@ -7,7 +7,6 @@ import VeiledereSide from './veiledere/veiledere-side';
 import TilbakemeldingFab from './components/tilbakemelding/tilbakemelding-fab';
 import {useFetchPortefoljeData} from './hooks/portefolje/use-fetch-portefolje-data';
 import Innholdslaster from './innholdslaster/innholdslaster';
-import TvungenStepper from './components/modal/tour-modal/tvungen-stepper';
 import {TilToppenKnapp} from './components/til-toppen-knapp/til-toppen-knapp';
 import './style.less';
 
@@ -26,7 +25,6 @@ function Routes() {
                     </Switch>
                     <TilbakemeldingFab />
                     <TilToppenKnapp />
-                    <TvungenStepper />
                 </Innholdslaster>
             </div>
         </div>
